Refactor for rebased code
