+++ embassy-mspm0/src/time_driver/tim.rs
@@ -7,7 +7,10 @@ use embassy_hal_internal::interrupt::InterruptExt;
 use embassy_time_driver::Driver;
 use embassy_time_queue_utils::Queue;
 use mspm0_metapac::interrupt;
-use mspm0_metapac::tim::vals::{Cm, Cvae, CxC, EvtCfg, PwrenKey, Repeat, ResetKey};
+use mspm0_metapac::tim::{
+    Tim,
+    vals::{Cm, Cvae, CxC, EvtCfg, PwrenKey, Repeat, ResetKey},
+};
 
 use crate::interrupt::typelevel::Interrupt;
 use crate::tim::SealedInstance;
